backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.0",
    "pyjwt>=2.0.0",
    "sqlmodel>=0.0.31",
]
